requirements.txt
flask==3.0.3
pdfplumber==0.11.4
Pillow==10.4.0
